# EasyTrufi.Api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backend EasyTrufi API",
    "description": "Documentación de la API de EasyTrufi - .NET 8",
    "contact": {
      "name": "Equipo de Desarrollo UCB",
      "email": "carlos.camacho.h@ucb.edu.bo"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Recupera una lista paginada de usuarios como objetos de transferencia de datos (DTO) según los filtros especificados.",
        "description": "Este método utiliza un mapeador para convertir las publicaciones recuperadas en DTO, que luego se \r\n            devuelven junto con la información de paginación. Si se produce un error durante el proceso, se devuelve un \r\n            código de estado 500 con los detalles del error.EasyTrufi.Api.Responses.ApiResponse`1",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Cedula",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna la lista de usuarios en formato DTO",
            "schema": {
              "$ref": "#/definitions/UserDTOIEnumerableApiResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Inserta un nuevo usuario en el sistema.",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "El objeto DTO que contiene los datos del usuario a insertar.",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos para la creación del usuario"
          }
        }
      }
    },
    "/api/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Recupera un usuario específico por su identificador único.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "El identificador único del usuario.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna el usuario solicitado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Actualiza los datos de un usuario existente.",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "El identificador único del usuario a actualizar.",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "description": "El objeto DTO que contiene los nuevos datos del usuario.",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario actualizado exitosamente"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Elimina un usuario del sistema.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "El identificador único del usuario a eliminar.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "Usuario eliminado exitosamente"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
    "/api/Users/{id}/verify-email": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Marca el correo del usuario como verificado (EmailVerified = true).",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Identificador del usuario a verificar.",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Correo verificado y usuario retornado",
            "schema": {
              "$ref": "#/definitions/UserResponseDTOApiResponse"
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Correo ya verificado",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Message": {
      "type": "object",
      "properties": {
        "type": {
          "description": "Tipo del mensaje (e.g., Información, Advertencia, Error)",
          "type": "string"
        },
        "description": {
          "description": "Descripción detallada del mensaje",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Pagination": {
      "type": "object",
      "properties": {
        "totalCount": {
          "format": "int32",
          "description": "Número total de elementos en la colección",
          "type": "integer"
        },
        "pageSize": {
          "format": "int32",
          "description": "Cantidad de elementos por página",
          "type": "integer"
        },
        "currentPage": {
          "format": "int32",
          "description": "Número de la página actual",
          "type": "integer"
        },
        "totalPages": {
          "format": "int32",
          "description": "Número total de páginas",
          "type": "integer"
        },
        "hasNextPage": {
          "description": "Indica si existe una página siguiente",
          "type": "boolean"
        },
        "hasPreviousPage": {
          "description": "Indica si existe una página anterior",
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
...
    "UserDTO": {
      "type": "object",
      "properties": {
        "cedula": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "dateOfBirth": {
          "format": "date",
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "passwordHash": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserDTOIEnumerableApiResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserDTO"
          }
        },
        "pagination": {
          "$ref": "#/definitions/Pagination"
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        }
      },
      "additionalProperties": false
    },
...
    "UserResponseDTO": {
      "type": "object",
      "properties": {
        "cedula": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "dateOfBirth": {
          "format": "date",
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserResponseDTOApiResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/UserResponseDTO"
        },
        "pagination": {
          "$ref": "#/definitions/Pagination"
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        }
      },
      "additionalProperties": false
    },
...
  }
}